feat: split actions by files
