+++ web/src/store/actions/index.ts
@@ -1,4 +1,6 @@
+export * from './actions';
 export * from './snippets';
 export * from './settings';
+export * from './editor';
 export * from './build';
-export * from './utils';
+export * from './ui';
